Switch GPU upload pipeline to fp32 origin-rebased gpu_sample_t
The 2b staging loop converted int64 ns to double seconds before upload;
the GPU shaders consumed that fp64 value through GL_ARB_gpu_shader_int64
and a packed-double SSBO layout. This is the slice-2c payoff: staging
produces gpu_sample_t directly, shaders go full fp32, and origin
rebasing keeps the rebased seconds value bounded by the visible window
even when absolute timestamps sit far off-axis.

Renderer:

- vbo_view_state_t now owns std::vector<gpu_sample_t> staging and an
  uploaded_t_origin_ns watermark. choose_origin_ns is invoked once per
  frame for main and once for preview; the two views own independent
  origins because their visible windows differ.
- Origin is part of the upload invalidation key. The stale-fallback
  predicate also requires uploaded_t_origin_ns == t_origin_ns, otherwise
  a frame whose origin moved into a new snap bucket would draw the old
  VBO contents against the new t_min_rel/t_max_rel uniforms and shift
  every sample by the bucket delta.
- One VAO per pipe with attributes pointing at gpu_sample_t at fixed
  offsets. AREA's instanced path binds locations 4/5 to the same VBO at
  offset sizeof(gpu_sample_t) so the shader sees (sample[i], sample[i+1])
  per instance. DOTS only declares locations 0/1, so 2/3 are explicitly
  disabled rather than left enabled with attribute pointers nothing reads.
- INVALID_SAMPLE_LAYOUT and the layout-validation block are gone — the
  fixed gpu_sample_t layout removes the failure mode they guarded.

Shaders:

- Drop GL_ARB_gpu_shader_int64.
- t_min, t_max, width, height, plot_area.frag's height all become float.
- Time vertex inputs are float in_x_rel / in_x0_rel / in_x1_rel.
- Sample_buffer SSBO is the fixed layout
  struct GpuSample { float t_rel; float y; float y_min; float y_max; };
  Drop uint raw[], packDouble2x32, u_sample_stride_uints,
  u_sample_x_offset_uints, u_sample_y_offset_uints.
- generic_rect, grid_quad, and text shaders are unchanged — they are
  not time shaders and consume absolute pixel/text coordinates.

Format_timestamp callbacks still receive absolute int64 ns; the rebase
is strictly a render-side detail. Axis labels, gridlines, and text
keep their absolute-timestamp pipeline so they cannot desync from data.

The Data_access_policy layout-byte fields and the Vertex_layout
infrastructure are no longer read by the renderer but remain in place;
batch 3 deletes them.

+++ shaders/plot_area.frag
@@ -2,7 +2,7 @@
 
 layout(location =  3) uniform float   v_min;
 layout(location =  4) uniform float   v_max;
-layout(location =  6) uniform double  height;
+layout(location =  6) uniform float   height;
 layout(location =  7) uniform float   y_offset;
 uniform float win_h;
 
@@ -20,7 +20,7 @@ void main(void)
 
     float vmin = v_min;
     float vmax = v_max;
-    float h    = float(height);
+    float h    = height;
 
     float vy = (1.0-(y / h)) * (vmax-vmin) + vmin;
     if (vy > vmax || vy < vmin) {
